+++ src/scripts/VariantEncumbranceImpl.ts
@@ -478,8 +478,8 @@ export const VariantEncumbranceImpl = {
 					debug(
 						`Is BackpackManager! Actor '${actorEntity.name}, Item '${item.name}' : Quantity = ${itemQuantity}, Weight = ${itemWeight}`
 					);
-					mapItemEncumbrance[<string>item.id] = (itemQuantity * itemWeight);
-					return weight + (itemQuantity * itemWeight);
+					mapItemEncumbrance[<string>item.id] = itemQuantity * itemWeight;
+					return weight + itemQuantity * itemWeight;
 				}
 
 				const isEquipped: boolean =
@@ -678,7 +678,7 @@ export const VariantEncumbranceImpl = {
 					}', Equipped '${isEquipped}', Proficient ${isProficient} :
              ${itemQuantity} * ${itemWeight} = ${appliedWeight} on total ${weight} => ${weight + appliedWeight}`
 				);
-				mapItemEncumbrance[<string>item.id] = (appliedWeight);
+				mapItemEncumbrance[<string>item.id] = appliedWeight;
 				return weight + appliedWeight;
 			}, 0);