+++ src/scripts/modules.ts
@@ -250,21 +250,23 @@ export const readyHooks = async () => {
 				// CUSTOMIZE INVENTORY
 				// ======================================================
 				const hideStandardWeightUnits = game.settings.get(CONSTANTS.MODULE_NAME, "hideStandardWeightUnits");
-				const replaceStandardWeightValue = game.settings.get(CONSTANTS.MODULE_NAME, "replaceStandardWeightValue");
+				const replaceStandardWeightValue = game.settings.get(
+					CONSTANTS.MODULE_NAME,
+					"replaceStandardWeightValue"
+				);
 				const isBulkEnable = game.settings.get(CONSTANTS.MODULE_NAME, "enableBulkSystem");
 
 				const listHeaders = htmlElement.find("li.items-header .item-weight");
 				for (const liHeaderB of listHeaders) {
 					//@ts-ignore
 					const liHeader = <JQuery<HTMLElement>>$(liHeaderB);
-					if(isBulkEnable) {
+					if (isBulkEnable) {
 						if (hideStandardWeightUnits) {
 							liHeader.text(`${getBulkLabel()}`);
 						} else {
 							liHeader.append(`<br/>${getBulkLabel()}`);
 						}
 					} else {
-
 					}
 				}
 
@@ -279,7 +281,12 @@ export const readyHooks = async () => {
 					VariantEncumbranceImpl.calculateEncumbrance(actorEntityTmp, inventoryItems, false, invPlusActive)
 				);
 				const encumbranceDataBulk = <EncumbranceBulkData>(
-					VariantEncumbranceBulkImpl.calculateEncumbrance(actorEntityTmp, inventoryItems, false, invPlusActive)
+					VariantEncumbranceBulkImpl.calculateEncumbrance(
+						actorEntityTmp,
+						inventoryItems,
+						false,
+						invPlusActive
+					)
 				);
 				const listItem = htmlElement.find("li.item .item-weight");
 				for (const liItemB of listItem) {
@@ -295,18 +302,24 @@ export const readyHooks = async () => {
 							.replace(/(\r\n|\n|\r)/gm, "")
 							.trim();
 						const currentTextB = currentText ? true : false;
-						
+
 						switch (sheetClass) {
 							case "dnd5e.Tidy5eSheet": {
-								if(replaceStandardWeightValue){
+								if (replaceStandardWeightValue) {
 									if (currentTextB) {
-										const weight =  encumbranceData.mapItemEncumbrance[<string>item.id]?.toNearest(0.1) ?? (quantity * getItemWeight(item)).toNearest(0.1) ?? 0;
+										const weight =
+											encumbranceData.mapItemEncumbrance[<string>item.id]?.toNearest(0.1) ??
+											(quantity * getItemWeight(item)).toNearest(0.1) ??
+											0;
 										const totalWeightS = `${weight} ${getWeightLabel()}`;
 										liItem.parent().find(".item-detail.item-weight").text(totalWeightS);
 									}
 								}
-								if(isBulkEnable) {
-									const bulk =  encumbranceDataBulk.mapItemEncumbrance[<string>item.id]?.toNearest(0.1) ?? (quantity * getItemBulk(item)).toNearest(0.1) ?? 0;
+								if (isBulkEnable) {
+									const bulk =
+										encumbranceDataBulk.mapItemEncumbrance[<string>item.id]?.toNearest(0.1) ??
+										(quantity * getItemBulk(item)).toNearest(0.1) ??
+										0;
 									const totalBulkS = `${bulk} ${getBulkLabel()}`;
 									if (hideStandardWeightUnits) {
 										if (currentTextB) {
@@ -314,22 +327,31 @@ export const readyHooks = async () => {
 										}
 									} else {
 										if (currentTextB) {
-											liItem.parent().find(".item-detail.item-weight").append(`<br/>${totalBulkS}`);
+											liItem
+												.parent()
+												.find(".item-detail.item-weight")
+												.append(`<br/>${totalBulkS}`);
 										}
 									}
 								}
 								break;
 							}
 							default: {
-								if(replaceStandardWeightValue){
+								if (replaceStandardWeightValue) {
 									if (currentTextB) {
-										const weight =  encumbranceData.mapItemEncumbrance[<string>item.id]?.toNearest(0.1) ?? (quantity * getItemWeight(item)).toNearest(0.1) ?? 0;
+										const weight =
+											encumbranceData.mapItemEncumbrance[<string>item.id]?.toNearest(0.1) ??
+											(quantity * getItemWeight(item)).toNearest(0.1) ??
+											0;
 										const totalWeightS = `${weight} ${getWeightLabel()}`;
 										liItem.parent().find(".item-detail.item-weight div").text(totalWeightS);
 									}
 								}
-								if(isBulkEnable) {
-									const bulk =  encumbranceDataBulk.mapItemEncumbrance[<string>item.id]?.toNearest(0.1) ?? (quantity * getItemBulk(item)).toNearest(0.1) ?? 0;
+								if (isBulkEnable) {
+									const bulk =
+										encumbranceDataBulk.mapItemEncumbrance[<string>item.id]?.toNearest(0.1) ??
+										(quantity * getItemBulk(item)).toNearest(0.1) ??
+										0;
 									const totalBulkS = `${bulk} ${getBulkLabel()}`;
 									if (hideStandardWeightUnits) {
 										if (currentTextB) {
@@ -337,7 +359,10 @@ export const readyHooks = async () => {
 										}
 									} else {
 										if (currentTextB) {
-											liItem.parent().find(".item-detail.item-weight div").append(`<br/>${totalBulkS}`);
+											liItem
+												.parent()
+												.find(".item-detail.item-weight div")
+												.append(`<br/>${totalBulkS}`);
 										}
 									}
 								}
@@ -1331,7 +1356,6 @@ const module = {
 		sheetClass: string
 	): Promise<void> {
 		if (game.settings.get(CONSTANTS.MODULE_NAME, "enableBulkSystem")) {
-
 			// ===============================
 			// CUSTOMIZE ENCUMBRANCE
 			// ===============================