+++ changelog.md
@@ -1,3 +1,8 @@
+### 0.12.7 
+
+- Add module settings 'replaceStandardWeightValue' by default is true 
+- bug fix backpack manager build weight calculation
+
 ### 0.12.6 
 
 - Bug fix : https://github.com/p4535992/foundryvtt-variant-encumbrance-dnd5e/issues/14